Code refactoring, moved Box<T> extensions to same file for clarity
